Refactor mapping tests to use CultureInfo.InvariantCulture for string conversions and enhance documentation
