+++ src/Core.Tests/Mapping/EnumerableMapperTests.cs
@@ -1,20 +1,27 @@
 ﻿using System.Collections.Generic;
+using System.Globalization;
 
 using Grove.Core.Mapping;
 
 using Moq;
 
 
-namespace Grove.Core.Tests;
+namespace Grove.Core.Tests.Mapping;
 
+/// <summary>
+///     Contains unit tests for the <see cref="EnumerableMapper{TFrom,TTo}" /> class.
+/// </summary>
 public class EnumerableMapperTests
 {
+    /// <summary>
+    ///     Tests that a collection of integers is correctly mapped to a collection of strings.
+    /// </summary>
     [Fact]
     public void MapsCollectionCorrectly()
     {
         // Arrange
         Mock<IMapper<int, string>> mockMapper = new();
-        mockMapper.Setup(m => m.Map(It.IsAny<int>())).Returns<int>(i => i.ToString());
+        mockMapper.Setup(m => m.Map(It.IsAny<int>())).Returns<int>(i => i.ToString(CultureInfo.InvariantCulture));
         EnumerableMapper<int, string> enumerableMapper = new(mockMapper.Object);
         List<int> input = new()
         {
@@ -37,13 +44,18 @@ public class EnumerableMapperTests
             result);
     }
 
+    /// <summary>
+    ///     Tests that an empty collection is correctly mapped to an empty collection.
+    /// </summary>
     [Fact]
     public void MapsEmptyCollection()
     {
         // Arrange
         Mock<IMapper<int, string>> mockMapper = new();
         EnumerableMapper<int, string> enumerableMapper = new(mockMapper.Object);
-        List<int> input = new();
+
+        // ReSharper disable once CollectionNeverUpdated.Local
+        List<int> input = [];
 
         // Act
         IEnumerable<string> result = enumerableMapper.Map(input);
@@ -52,12 +64,16 @@ public class EnumerableMapperTests
         Assert.Empty(result);
     }
 
+    /// <summary>
+    ///     Tests that a collection containing nullable integers, including null elements,
+    ///     is correctly mapped to a collection of strings.
+    /// </summary>
     [Fact]
     public void MapsCollectionWithNullElements()
     {
         // Arrange
         Mock<IMapper<int?, string>> mockMapper = new();
-        mockMapper.Setup(m => m.Map(It.IsAny<int?>())).Returns<int?>(i => i?.ToString() ?? "null");
+        mockMapper.Setup(m => m.Map(It.IsAny<int?>())).Returns<int?>(i => i?.ToString(CultureInfo.InvariantCulture) ?? "null");
         EnumerableMapper<int?, string> enumerableMapper = new(mockMapper.Object);
         List<int?> input = new()
         {
