+++ src/Core.Tests/Mapping/MappingRegistrationsTests.cs
@@ -1,12 +1,20 @@
-﻿using Grove.Core.Mapping;
+﻿using System.Globalization;
+
+using Grove.Core.Mapping;
 
 using Microsoft.Extensions.DependencyInjection;
 
 
 namespace Grove.Core.Tests.Mapping;
 
+/// <summary>
+/// Contains unit tests for verifying the registration of mappers in the service collection.
+/// </summary>
 public class MappingRegistrationsTests
 {
+    /// <summary>
+    /// Tests if a single mapper is correctly added to the service collection.
+    /// </summary>
     [Fact]
     public void AddsMapperToServiceCollection()
     {
@@ -18,6 +26,9 @@ public class MappingRegistrationsTests
         Assert.IsType<MockMapper>(mapper);
     }
 
+    /// <summary>
+    /// Tests if an IEnumerable mapper is correctly added to the service collection.
+    /// </summary>
     [Fact]
     public void AddsIEnumerableMapperToServiceCollection()
     {
@@ -30,6 +41,9 @@ public class MappingRegistrationsTests
         Assert.IsType<EnumerableMapper<int, string>>(mapper);
     }
 
+    /// <summary>
+    /// Tests if an IAsyncEnumerable mapper is correctly added to the service collection.
+    /// </summary>
     [Fact]
     public void AddsIAsyncEnumerableMapperToServiceCollection()
     {
@@ -42,11 +56,20 @@ public class MappingRegistrationsTests
         Assert.IsType<AsyncEnumerableMapper<int, string>>(mapper);
     }
 
+    /// <summary>
+    /// A mock implementation of the <see cref="IMapper{TFrom, TTo}"/> interface
+    /// for testing purposes.
+    /// </summary>
     private class MockMapper : IMapper<int, string>
     {
+        /// <summary>
+        /// Maps an integer to its string representation.
+        /// </summary>
+        /// <param name="source">The integer to map.</param>
+        /// <returns>The string representation of the integer.</returns>
         public string Map(
             int source
         ) =>
-            source.ToString();
+            source.ToString(CultureInfo.InvariantCulture);
     }
 }
